chore: remove price context
